mlops/mlops_capstone/frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = "the frontend application of the project"
authors = ["Alex Beckwith <57322567+witzbeck@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "*"
streamlit = "*"
Pillow = "*"
matplotlib = "*"
requests = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
